refactor: change attribute filter component from a dropdown to a button
