+++ packages/ui/src/index.ts
@@ -37,7 +37,7 @@ export {
 	type ToolbarProps,
 
 	// & individual parts for integrators to build their own
-	AttributeFilterDropdown,
+	AttributeFilterButton,
 	CollapseAllButton,
 	type CollapseAllButtonProps,
 	DictionaryDownloadButton,
