+++ packages/ui/src/viewer-table/Toolbar/index.ts
@@ -1,4 +1,5 @@
-export { default as AttributeFilterDropdown } from './AttributeFilterDropdown.js';
+export { default as AttributeFilterButton } from './AttributeFilterButton.js';
+export { default as FilterRows, type FilterRowsProps } from './FilterRows.js';
 export { default as CollapseAllButton, type CollapseAllButtonProps } from './CollapseAllButton.js';
 export { default as DiagramViewButton } from './DiagramViewButton.js';
 export { default as DictionaryDownloadButton, type DictionaryDownloadButtonProps } from './DictionaryDownloadButton.js';
